% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btrany.r
\name{btrany}
\alias{btrany}
\title{Back transform the y values to the original scale}
\usage{
btrany(y, tau = 100, type = 1, a = tau, b = 1, c = 0)
}
\arguments{
\item{y}{A vector of transformed y values, y must be in interval [0,1].}

\item{tau}{A constant denoting the upper bound of the observed data.}

\item{type}{Type of the transformation, 1: Gamma type, 2: Weibull type, and other values: uniform}

\item{a}{Parameter a in the transformation.}

\item{b}{Parameter b in the transformation.}

\item{c}{Parameter c in the transformation.}
}
\value{
Returns the back-transformed values in the original scale.
}
\description{
This function back transforms the y values to the original scale.
}
\examples{
y=runif(100)
btrany(y=y,tau=10,type=1,a=10,b=2,c=0.5)

}
