\name{PWEALL-package}
\alias{PWEALL-package}
\alias{PWEALL}
\docType{package}
\title{
\packageTitle{PWEALL}
}
\description{
\packageDescription{PWEALL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PWEALL}
\packageIndices{PWEALL}
}
\author{
\packageAuthor{PWEALL}

Maintainer: \packageMaintainer{PWEALL}
}
\references{
Luo et al. (2017)
}
\keyword{piecewise exponential}
\keyword{various functions}
\seealso{
Luo et al. (2017)
}
