\name{rmstsim}
\alias{rmstsim}
\title{simulating the restricted mean survival times}
\description{This will simulate the test statistics accouting for staggered entry, delayed treatment effect,  treatment crossover and loss to follow-up.}
\usage{
rmstsim(tcut=c(1,2),tstudy=tcut+0.2,taur=1.2,
        u=c(1/taur,1/taur),ut=c(taur/2,taur),pi1=0.5,
        rate11=c(1,0.5),rate21=rate11,rate31=c(0.7,0.4),
        rate41=rate21,rate51=rate21,ratec1=c(0.5,0.6),
        rate10=rate11,rate20=rate10,rate30=rate31,
        rate40=rate20,rate50=rate20,ratec0=c(0.6,0.5),
        tchange=c(0,1),type1=1,type0=1,rp21=0.5,rp20=0.5,
        n=1000,rn=200,eps=1.0E-08)
}
\arguments{
  \item{tcut}{a vector of time points at which rmst are calculated}
  \item{tstudy}{a vector of study time points, should be the same length as \code{tcut} and should be not less than \code{tcut} element-wise}
  \item{taur}{Recruitment time}
  \item{u}{Piecewise constant recuitment rate}
  \item{ut}{Recruitment intervals}
  \item{pi1}{Allocation probability for the treatment group}
  \item{rate11}{Hazard before crossover for the treatment group}
  \item{rate21}{Hazard after crossover for the treatment group}
  \item{rate31}{Hazard for time to crossover for the treatment group}
  \item{rate41}{Hazard after crossover for the treatment group for complex case}
  \item{rate51}{Hazard after crossover for the treatment group for complex case}
  \item{ratec1}{Hazard for time to censoring for the treatment group}
  \item{rate10}{Hazard before crossover for the control group}
  \item{rate20}{Hazard after crossover for the control group}
  \item{rate30}{Hazard for time to crossover for the control group}
  \item{rate40}{Hazard after crossover for the control group for complex case}
  \item{rate50}{Hazard after crossover for the control group for complex case}
  \item{ratec0}{Hazard for time to censoring for the control group}
  \item{tchange}{A strictly increasing sequence of time points at which the event rates changes. The first element of tchange                 must be zero. It must have the same length as \code{rate11}, \code{rate21}, \code{rate31}, etc.}
  \item{type1}{Type of crossover in the treatment group}
  \item{type0}{Type of crossover in the control group}
  \item{rp21}{re-randomization prob in the treatment group}
  \item{rp20}{re-randomization prob in the control group}
  \item{n}{number of subjects}
  \item{rn}{number of simulations}
  \item{eps}{tolerence for comparing event times}
}
\details{
The hazard functions corresponding to \code{rate11},...,\code{rate51},\code{ratec1}, \code{rate10},...,\code{rate50},\code{ratec0} are all piecewise constant function taking the form  \eqn{\lambda(t)=\sum_{j=1}^m \lambda_j I(t_{j-1}\le t<t_j)}, where \eqn{\lambda_1,\ldots,\lambda_m} are the corresponding elements of the rates and \eqn{t_0,\ldots,t_{m-1}} are the corresponding elements of \cite{tchange}, \eqn{t_m=\infty}. Note that all the rates must have the same \code{tchange}.
}
\value{\item{outr}{test statistics at each pair of \code{tcut} and \code{tstudy} in column and each simulation run in row}
}
\references{
Luo et al. (2018) Design and monitoring of survival trials in complex scenarios, Statistics in Medicine <doi: https://doi.org/10.1002/sim.7975>.
}
\author{Xiaodong Luo}
\note{Version 1.0 (7/19/2016)}
\seealso{\code{\link{pwe}},\code{\link{rpwe}},\code{\link{qpwe}},\code{\link{ovbeta}}}
\examples{
tcuta<-c(2,3)
taur<-1.2
u<-c(1/taur,1/taur)
ut<-c(taur/2,taur)
r11<-c(1,0.5)
r21<-c(0.5,0.8)
r31<-c(0.7,0.4)
r41<-r51<-r21
rc1<-c(0.5,0.6)
r10<-c(1.5,0.7)
r20<-c(0.5,1)
r30<-c(0.3,0.4)
r40<-r50<-r20
rc0<-c(0.2,0.4)
ar<-rmstsim(tcut=tcuta,tstudy=tcuta+0.1,taur=taur,u=u,ut=ut,pi1=0.5,
            rate11=r11,rate21=r21,rate31=r31,rate41=r41,rate51=r51,ratec1=rc1,
            rate10=r10,rate20=r20,rate30=r30,rate40=r40,rate50=r50,ratec0=rc0,
            tchange=c(0,1),type1=1,type0=1,
            n=300,rn=200)
##Empirical power
apply(ar$outr>1.96,2,mean)        
}
\keyword{piecewise exponential}
\keyword{overall hazard ratio}
\keyword{piecewise uniform}
\keyword{delayed treatment effect}
\keyword{treatment crossover}
\keyword{simulation}

