% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.to.logistic.r
\name{surv.to.logistic}
\alias{surv.to.logistic}
\title{Transform the two-sample survival data into logistic regression data}
\usage{
surv.to.logistic(y, d, z, tcut = max(y[d == 1]))
}
\arguments{
\item{y}{A vector of observed times.}

\item{d}{A vector of event indicators with 1 denoting event and 0 denoting censoring.}

\item{z}{A vector of group indicators with 1 being the treatment and 0 the control.}

\item{tcut}{A constant as the cut-point.}
}
\value{
A dataframe for logistic regression with z denoting 1/0 outcome, y the (univariate) covariate and yset the offset.
}
\description{
The following function transforms the two-sample survival data into a logistic regression model with an offset.
}
\examples{
n1=100;n2=100;n=n1+n0
zz=rep(0,n);zz[1:n1]=1
tt=c(rexp(n1)/0.1,rexp(n0)/0.2)
cc=c(rexp(n1)/0.05,rexp(n0)/0.06)
yy=pmin(tt,cc)
dd=(tt<=yy)
df=surv.to.logistic(y=yy,d=dd,z=zz,tcut=5)$adata
 
}
