% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbase.r
\name{tbase}
\alias{tbase}
\title{Generate Truncated Spline Base}
\usage{
tbase(x, iknots, lk, hk, deg = 3)
}
\arguments{
\item{x}{A vector.}

\item{iknots}{A vector of inner knots.}

\item{lk}{lower boundary knot.}

\item{hk}{higher boundary knot.}

\item{deg}{degree of the splines, default is 3.}
}
\value{
The truncated spline base.
}
\description{
Generate Truncated Spline Base
}
\examples{
tbase(x=c(0.25,0.35,0.5,0.7),iknots=c(0.2,0.4,0.6,0.8),lk=0,hk=1,deg=3)

}
