% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trany.r
\name{trany}
\alias{trany}
\title{Transform the observed survival times into [0,1]}
\usage{
trany(y, tau = max(y) + 0.1, type = 1, a = tau, b = 1, c = 0)
}
\arguments{
\item{y}{A vector of observed times to be transformed.}

\item{tau}{A constant denoting the upper bound of the observed data.}

\item{type}{Type of the transformation, 1: Gamma type, 2: Weibull type, and other values: uniform}

\item{a}{Parameter a in the transformation.}

\item{b}{Parameter b in the transformation.}

\item{c}{Parameter c in the transformation.}
}
\value{
Returns the transformed values in [0,1].
}
\description{
This function transforms the observed survival times into [0,1].
}
\examples{
y=rexp(100)
trany(y=y,tau=10,type=1,a=10,b=2,c=0.5)

}
